// packages/shared/src/types/physics.ts
export type PxVec3 = PhysX.PxVec3;
export type PxQuat = PhysX.PxQuat;
export type PxTransform = PhysX.PxTransform;

// Actor types
export type PxActor = PhysX.PxActor;
export type PxRigidActor = PhysX.PxRigidActor;
export type PxRigidBody = PhysX.PxRigidBody;
export type PxRigidDynamic = PhysX.PxRigidDynamic;
// ...
export type PxFilterData = PhysX.PxFilterData;
// ...
export type PxController = PhysX.PxController;
export type PxControllerManager = PhysX.PxControllerManager;
export type PxControllerDesc = PhysX.PxControllerDesc;
export type PxCapsuleControllerDesc = PhysX.PxCapsuleControllerDesc;
export type PxBoxControllerDesc = PhysX.PxBoxControllerDesc;
export type PxControllerFilters = PhysX.PxControllerFilters;
export type PxControllerCollisionFlags = PhysX.PxControllerCollisionFlags;
export type PxExtendedVec3 = PhysX.PxExtendedVec3;
export type PxObstacleContext = PhysX.PxObstacleContext;
// ...
export interface PhysXController extends PxController {
  getActor(): PxRigidDynamic;
  setFootPosition(position: PxExtendedVec3): boolean;
  move(
    disp: PxVec3,
    minDist: number,
    elapsedTime: number,
    filters: PxControllerFilters,
    obstacles?: PxObstacleContext,
  ): PxControllerCollisionFlags;
}
// ...
export interface ActorHandle {
  move: (matrix: THREE.Matrix4) => void;
  snap: (pose: PxTransform) => void;
  destroy: () => void;
}